Adding Titles, Wrong Title Fix
What's changed:
+ Publisher's Home
+ Ability to Add Titles
+ Fixed: Due to Cookies wrong title is displayed in title and publisher page

+++ handler/account.logout.php
@@ -3,6 +3,12 @@
 require("../load.php");
 require("../sql/account.php");
 
+if ($loggedin == false) {
+    header("Location: " . config("url") . "account/login");
+    $error = false;
+    die("You're not even logged in? If you are, contact the developers - this is a bug! (Also, you are ignoring headers, this isn't a good sign!)");
+}
+
 // Removing token from Database and destroy entire session and so on
 $token = clean(mysqli_real_escape_string($conn, $_COOKIE[config("cookie") . "_session"]));
 $conn->query("DELETE FROM `sessions` WHERE `token`='$token'");
